reorganize graph modifier
