+++ .changeset/purple-stingrays-clap.md
@@ -1,0 +1,7 @@
+---
+'@finos/legend-extension-dsl-diagram': patch
+'@finos/legend-extension-dsl-text': patch
+'@finos/legend-graph': patch
+'@finos/legend-query': patch
+'@finos/legend-studio': patch
+---
