+++ packages/legend-studio/src/components/editor/edit-panel/connection-editor/FlatDataConnectionEditor.tsx
@@ -16,7 +16,7 @@
 
 import { observer } from 'mobx-react-lite';
 import type { FlatDataConnectionValueState } from '../../../../stores/editor-state/element-editor-state/connection/ConnectionEditorState';
-import { flatData_setUrl } from '../../../../stores/graphModifier/GraphModifierHelper';
+import { flatDataConnection_setUrl } from '../../../../stores/graphModifier/StoreFlatData_GraphModifierHelper';
 
 export const FlatDataConnectionEditor = observer(
   (props: {
@@ -27,7 +27,7 @@ export const FlatDataConnectionEditor = observer(
     const connection = connectionValueState.connection;
     // url
     const changeUrl: React.ChangeEventHandler<HTMLTextAreaElement> = (event) =>
-      flatData_setUrl(connection, event.target.value);
+      flatDataConnection_setUrl(connection, event.target.value);
     return (
       <div className="panel__content__form">
         <div className="panel__content__form__section">
