+++ packages/legend-studio/src/components/editor/edit-panel/mapping-editor/MappingExecutionBuilder.tsx
@@ -76,10 +76,8 @@ import {
 } from '@finos/legend-graph';
 import { StudioTextInputEditor } from '../../../shared/StudioTextInputEditor';
 import type { DSLMapping_LegendStudioPlugin_Extension } from '../../../../stores/DSLMapping_LegendStudioPlugin_Extension';
-import {
-  flatData_setData,
-  objectInputData_setData,
-} from '../../../../stores/graphModifier/GraphModifierHelper';
+import { objectInputData_setData } from '../../../../stores/graphModifier/DSLMapping_GraphModifierHelper';
+import { flatData_setData } from '../../../../stores/graphModifier/StoreFlatData_GraphModifierHelper';
 import {
   relationalInputData_setData,
   relationalInputData_setInputType,
