+++ packages/legend-studio/src/index.ts
@@ -49,7 +49,7 @@ export { ExplorerTreeRootPackageLabel } from './stores/ExplorerTreeState';
 export * from './stores/DSLMapping_LegendStudioPlugin_Extension';
 export * from './stores/DSLService_LegendStudioPlugin_Extension';
 export * from './stores/graphModifier/DomainGraphModifierHelper';
-export * from './stores/graphModifier/GraphModifierHelper';
+export * from './stores/graphModifier/DSLGeneration_GraphModifierHelper';
 export * from './stores/graphModifier/DSLService_GraphModifierHelper';
 
 // components
