+++ packages/legend-studio/src/stores/NewElementState.tsx
@@ -75,12 +75,14 @@ import {
 import type { DSLMapping_LegendStudioPlugin_Extension } from './DSLMapping_LegendStudioPlugin_Extension';
 import { package_addElement } from './graphModifier/DomainGraphModifierHelper';
 import {
+  packageableConnection_setConnectionValue,
+  runtime_addMapping,
+} from './graphModifier/DSLMapping_GraphModifierHelper';
+import {
   fileGeneration_setScopeElements,
   fileGeneration_setType,
   generationSpecification_addGenerationElement,
-  packageableConnection_setConnectionValue,
-  runtime_addMapping,
-} from './graphModifier/GraphModifierHelper';
+} from './graphModifier/DSLGeneration_GraphModifierHelper';
 import {
   service_initNewService,
   service_setExecution,
