+++ packages/legend-studio/src/stores/graphModifier/ValueSpecificationGraphModifierHelper.ts
@@ -1,0 +1,40 @@
+/**
+ * Copyright (c) 2020-present, Goldman Sachs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import type {
+  Multiplicity,
+  RawVariableExpression,
+  Type,
+} from '@finos/legend-graph';
+import { action } from 'mobx';
+
+// --------------------------------------------- Raw Value Specification -------------------------------------
+
+export const rawVariableExpression_setName = action(
+  (rV: RawVariableExpression, value: string): void => {
+    rV.name = value;
+  },
+);
+export const rawVariableExpression_setType = action(
+  (rV: RawVariableExpression, value: Type): void => {
+    rV.type.value = value;
+  },
+);
+export const rawVariableExpression_setMultiplicity = action(
+  (rV: RawVariableExpression, value: Multiplicity): void => {
+    rV.multiplicity = value;
+  },
+);
